pyproject.toml
[tool.poetry]
name = "app-for-analyz-bank-transactions"
version = "0.1.0"
description = ""
authors = ["Evgeny Safonov <jeka-luda@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
openpyxl = "^3.1.2"
requests = "^2.31.0"
xlrd = "^2.0.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.12.0"

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.black]
# максимальная длина строки
line-length = 119
# файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # исключить несколько общих каталогов в
    | \.git          # корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # также отдельно исключить файл с именем foo.py в
                     # корне проекта
)
'''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
